+++ Mydemenageur.API/Models/Housing/HousingRegisterModel.cs
@@ -32,7 +32,7 @@ namespace Mydemenageur.API.Models.Housing
         /// The housing surface
         /// </summary>
         /// <example>36</example>
-        public float Surface { get; set; }
+        public double Surface { get; set; }
         [Required]
         /// <summary>
         /// The housing Address
