+++ Mydemenageur.API/Models/Housing/HousingUpdateModel.cs
@@ -33,7 +33,7 @@ namespace Mydemenageur.API.Models.Housing
         /// The housing surface
         /// </summary>
         /// <example>36</example>
-        public float Surface { get; set; }
+        public double Surface { get; set; }
         [Required]
         /// <summary>
         /// The housing Address
