+++ Mydemenageur.API/Models/Vehicule/VehiculesUpdateModel.cs
@@ -22,7 +22,7 @@ namespace Mydemenageur.API.Models.Vehicule
         /// <summary>
         /// PTAC of tarpaulin ptac
         /// </summary>
-        public float PTAC_TarpaulinVehicule { get; set; }
+        public double PTAC_TarpaulinVehicule { get; set; }
         [Required]
         /// <summary>
         /// If the society have hard wall vehicule
@@ -32,7 +32,7 @@ namespace Mydemenageur.API.Models.Vehicule
         /// <summary>
         /// PTAC of hard wall ptac
         /// </summary>
-        public float PTAC_HardWallVehicule { get; set; }
+        public double PTAC_HardWallVehicule { get; set; }
         [Required]
         /// <summary>
         /// If he can transport horse
@@ -47,6 +47,6 @@ namespace Mydemenageur.API.Models.Vehicule
         /// <summary>
         /// Total capacity of all vehicule in the society
         /// </summary>
-        public float TotalCapacity { get; set; }
+        public double TotalCapacity { get; set; }
     }
 }
